Added tests

+++ src/instructions/jr.c
@@ -25,7 +25,7 @@ void  cpu_jump_rel(struct cpu_state *state, reg_t addr)
 	// by expanding both to 32 bits, performing the
 	// subtraction and casting back to 16 bits.
 	//TODO:Clean up to and make sure it is correct.
-	int32_t t = (int32_t)*(int8_t*)&addr;
-	int32_t pc = state->pc;
+	int16_t t = *(int8_t*)&addr;
+	int16_t pc = state->pc;
 	state->pc = pc + t;
 }
